Move new_revision contextmanager from tracim.model.data to tracim.model
